+++ examples/src/test/scala/commbank/coppersmith/examples/userguide/DirectorFeaturesSpec.scala
@@ -19,7 +19,7 @@ import org.scalacheck.{Gen, Arbitrary}
 
 import au.com.cba.omnia.thermometer.hive.ThermometerHiveSpec
 
-import commbank.coppersmith.api._
+import commbank.coppersmith.api._, Coppersmith._
 import commbank.coppersmith.examples.thrift.{Rating, Movie}
 
 object DirectorFeaturesSpec extends ThermometerHiveSpec {